feat: `all purge` command
